+++ src/App.tsx
@@ -12,6 +12,7 @@ import ReviewCenter from './pages/ReviewCenter'
 import WeakWords from './pages/WeakWords'
 import Scenes from './pages/Scenes'
 import LearnReport from './pages/LearnReport'
+import PlanPage from './pages/PlanPage'
 import SceneDetail from './pages/SceneDetail'
 import Camera from './pages/Camera'
 import AIChat from './pages/AIChat'
@@ -98,6 +99,7 @@ function App() {
         <Route path="scenes/:id" element={<SceneDetail />} />
         <Route path="camera" element={<Camera />} />
         <Route path="chat" element={<AIChat />} />
+        <Route path="plan" element={<PlanPage />} />
         <Route path="report" element={<LearnReport />} />
         <Route path="settings" element={<Settings />} />
         <Route path="*" element={<Navigate to="/" replace />} />